Squash before merging
