+++ oap-data-source/arrow/src/main/scala/org/apache/spark/sql/execution/datasources/v2/arrow/ArrowUtils.scala
@@ -20,10 +20,14 @@ import java.net.URI
 import java.util.TimeZone
 
 import org.apache.arrow.dataset.file.{FileSystem, SingleFileDataSourceDiscovery}
+import org.apache.arrow.vector.VectorSchemaRoot
 import org.apache.arrow.vector.types.pojo.Schema
 import org.apache.hadoop.fs.FileStatus
+import org.apache.spark.sql.catalyst.InternalRow
+import org.apache.spark.sql.execution.vectorized.{ArrowWritableColumnVector, ColumnVectorUtils, OnHeapColumnVector}
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.util.CaseInsensitiveStringMap
+import org.apache.spark.sql.vectorized.ColumnarBatch
 
 import scala.collection.JavaConverters._
 
@@ -68,6 +72,21 @@ object ArrowUtils {
   def toArrowSchema(t: StructType): Schema = {
     // fixme this might be platform dependent
     org.apache.spark.sql.util.ArrowUtils.toArrowSchema(t, TimeZone.getDefault.getID)
+  }
+
+  def loadVsr(vsr: VectorSchemaRoot, partitionValues: InternalRow, partitionSchema: StructType): ColumnarBatch = {
+    val fvs = vsr.getFieldVectors
+
+    val rowCount = vsr.getRowCount
+    val vectors = ArrowWritableColumnVector.loadColumns(rowCount, fvs)
+    val partitionColumns = OnHeapColumnVector.allocateColumns(rowCount, partitionSchema)
+    (0 until partitionColumns.length).foreach(i => {
+      ColumnVectorUtils.populate(partitionColumns(i), partitionValues, i)
+      partitionColumns(i).setIsConstant()
+    })
+
+    val batch = new ColumnarBatch(vectors ++ partitionColumns, rowCount, new Array[Long](5))
+    batch
   }
 
   private def getFormat(
